Parse a "block" of characters into a Digit
We'd like to be able to parse a string of 3x3 "ASCII-art" representations of
digits into entries representing account numbers. This is a first step toward
that; a `DigitParser` takes a single 3x3 "block" of characters and parses it
into either a `Digit` or `IllegibleDigit` object.

While we don't technically *need* an `IllegibleDigit` object yet (User Story #3
references illegible digits, but I usually don't like to write more code than is
necessary Right Now), I'd argue that including it already clarifies the code.

If the raw block passed into the `DigitParser::BLOCKS` hash doesn't match any of
the pre-defined entries, the hash lookup will return `nil`. We could ignore the
failed lookup and propagate `nil` as the value of the new `Digit`, but that (1)
taints every `Digit` object, since some of them may now return `nil` values,
which we'll have to check for everywhere, and (2) doesn't make it clear that
we've explicitly considered this situation and chosen to handle it.