Parse a collection of lines into a single Entry
Our next goal will be to parse a file of entries into a collection of Entry
objects. Meanwhile, this allows us to parse an array of three lines into a
single `Entry`.

An `Entry` is (so far) just a simple value object that containing a collection
of `Digit`s.

We're being intentionally strict about what we accept here; trying to parse an
`Entry` with an incorrect number of lines or characters raises an exception. If
this were for "real" work I'd try to get some clarification on how to handle
failure. In the absence of guidance, though, I'd rather just fail fast and hard.