Extract Entry objects from a plain text file
This allows instantiation of a collection of `Entry` objects with:

    file = EntryFile.new(File.open("path/to/file.txt"))

    file.entries # => An Array of Entry objects

Note that the constructor takes a `File` object, not a path. This lets us pass
in anything that acts like a `File`. That's nice for testing, I think, since we
don't need to muck about with the file system.

This closes "User Story #1."