Determine whether an Entry's checksum is valid
This adds a public `Entry#valid?` method, which returns true iff every digit in
the entry is legible and the entry's checksum is a multiple of 11.

I've chosen to keep `Entry#checksum` private for now. It's an implementation
detail at the moment, but we might want to expose it eventually. Keeping it
private also ensures that we can't call it without ascertaining that the entry
contains no illegible digits.

This also removes the `source_block` arguments on the `Digit` and
`IllegibleDigit` constructors. I was kinda thinking that I might to keep that
information around for "User Story #4," but (1) handling it was uglifying the
tests, and (2) we can always add it back in. YAGNI, amirite?

This closes "User Story #2."